Rename [#46] formatDate 유틸함수 파일명 변경
